requirements.txt
tensorflow
numpy
pandas
matplotlib
scikit-learn
gradio
python-dotenv
twilio
pillow
